drm/i915: add empty out_fence support
Enable to userspace to save a roundtrip to the GPU side. Instead of
getting the out_fence fd back when execbuf returns we create an empty
sync_file fd and send it to i915 and DRM/KMS at the same moment without
waiting for the round trip.

i915 receives the empty sync file and bounds it with its internal
request's fence. Then KMS (or any other driver) can get the fence and
wait for the job to complete.

Signed-off-by: Gustavo Padovan <gustavo.padovan@collabora.com>