drm/atomic: Introduce pending properties for plane updates
For async plane update there is a need to know if the current pending
atomic state is touching the plane we are trying to modify
asynchronously.

Thus we introduce the pending member on drm_plane_state to take account
of that. Once the state is swapped there is currently no way know which
planes are part of the new update and which aren't.

plane->state->pending is set to true on drm_atomic_helper_swap_state()
and back to false on drm_atomic_helper_commit_hw_done().

Signed-off-by: Gustavo Padovan <gustavo.padovan@collabora.com>